Create RaspiMicrogreensController
This is a summer project I did while studying python. 

How it works:
Raspberry Pi sends signals to an Arduino to turn relays on and off at a certain point in time. I could have used a raspberry pi relay directly but I already have an Arduino sitting around.
The light will turn on at around 7:00 in the morning and around 10:00pm. This can be modified in the code.

Things that need improvement:
The timer for the fan should be changed into something scale-able. I have not figured that out yet. Plan is for the fan to run for 10 minutes every hour.
I am also planning on using the DHT11 temperature and relative humidity sensor so control the humidity indoors to create a perfect growing condition for the plants. Right now this is just being displayed.

Implementation:
The python program runs as soon as raspi is turned on. I followed the instruction in the below website.
https://learn.sparkfun.com/tutorials/how-to-run-a-raspberry-pi-program-on-startup/all